DEP-3877: Added SA DAv3 page
