+++ app/config/AppConfig.scala
@@ -39,6 +39,7 @@ class AppConfig @Inject()(config: Configuration) {
   val showVATCUI: Boolean = config.getOptional[Boolean]("features.showVATCUI").getOrElse(false)
   val showCTCUI: Boolean = config.getOptional[Boolean]("features.showCTCUI").getOrElse(false)
   val showCISCUI: Boolean = config.getOptional[Boolean]("features.showCISCUI").getOrElse(false)
+  val showSACUI: Boolean = config.getOptional[Boolean]("features.showSACUI").getOrElse(false)
 
   // Used in wrapper
   val analyticsToken: String = config.get[String]("google-analytics.token")
