+++ app/controllers/CiapiController/CiapiController.scala
@@ -20,7 +20,7 @@ import config.AppConfig
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
 import views.html.CIAPIViews._
-import views.html.CUIViews.{childBenefitCUIDAv2View,ConstructionIndustrySchemeCUIDAv2View}
+import views.html.CUIViews.{childBenefitCUIDAv2View,ConstructionIndustrySchemeCUIDAv2View, SelfAssessmentCUIDAv2View}
 import javax.inject.{Inject, Singleton}
 
 import scala.concurrent.Future
@@ -35,7 +35,9 @@ class CiapiController @Inject()(appConfig: AppConfig,
                                 childBenefitCUIDAv2View: childBenefitCUIDAv2View,
                                 childBenefitCUIView: ChildBenefitCUIView,
                                 constructionIndustrySchemeCUIDAv2View: ConstructionIndustrySchemeCUIDAv2View,
-                                constructionIndustrySchemeCUIView: ConstructionIndustrySchemeCUIView) extends FrontendController(mcc) {
+                                constructionIndustrySchemeCUIView: ConstructionIndustrySchemeCUIView,
+                                selfAssessmentCUIDAv2View: SelfAssessmentCUIDAv2View,
+                                selfAssessmentCUIView: SelfAssessmentCUIView) extends FrontendController(mcc) {
 
   implicit val config: AppConfig = appConfig
 
@@ -80,6 +82,14 @@ class CiapiController @Inject()(appConfig: AppConfig,
       Future.successful(Ok(constructionIndustrySchemeCUIView()))
     } else {
       Future.successful(Ok(constructionIndustrySchemeCUIDAv2View()))
+    }
+  }
+
+  def selfAssessment: Action[AnyContent] = Action.async { implicit request =>
+    if (config.showSACUI) {
+      Future.successful(Ok(selfAssessmentCUIView()))
+    } else {
+      Future.successful(Ok(selfAssessmentCUIDAv2View()))
     }
   }
 }
