+++ app/controllers/CuiController/CuiController.scala
@@ -26,15 +26,10 @@ import scala.concurrent.Future
 @Singleton
 class CuiController @Inject()(appConfig: AppConfig,
                               mcc: MessagesControllerComponents,
-                              selfAssessmentCUIView: SelfAssessmentCUIView,
                               onlineServicesHelpdeskCUIView: OnlineServicesHelpdeskCUIView,
                               employerEnquiriesCUIView: EmployerEnquiriesCUIView) extends FrontendController(mcc) {
 
   implicit val config: AppConfig = appConfig
-
-  def selfAssessment: Action[AnyContent] = Action.async { implicit request =>
-      Future.successful(Ok(selfAssessmentCUIView()))
-  }
 
   def onlineServicesHelpdesk: Action[AnyContent] = Action.async { implicit request =>
       Future.successful(Ok(onlineServicesHelpdeskCUIView()))
