+++ conf/app.routes
@@ -36,7 +36,6 @@ GET         /webchat/payment-problems                                           
 
 GET         /accessibility-statement                                                    controllers.AccessibilityStatementController.accessibility(userAction: Option[String])
 
-GET         /chat/self-assessment                                                       controllers.CuiController.CuiController.selfAssessment
 GET         /chat/online-services-helpdesk                                              controllers.CuiController.CuiController.onlineServicesHelpdesk
 GET         /chat/employer-enquiries                                                    controllers.CuiController.CuiController.employerEnquiries
 
@@ -46,3 +45,4 @@ GET			/chat/vat-online															controllers.CiapiController.CiapiController
 GET         /chat/corporation-tax                                                       controllers.CiapiController.CiapiController.corporationTax
 GET         /chat/child-benefit                                                         controllers.CiapiController.CiapiController.childBenefit
 GET         /chat/construction-industry-scheme                                          controllers.CiapiController.CiapiController.constructionIndustryScheme
+GET         /chat/self-assessment                                                       controllers.CiapiController.CiapiController.selfAssessment
