+++ conf/application.conf
@@ -160,4 +160,5 @@ features {
   showVATCUI = true
   showCTCUI = true
   showCISCUI = true
+  showSACUI = true
 }
