+++ test/views/html/pages/includes/cuiSASidebarSpec.scala
@@ -19,12 +19,12 @@ package views.html.pages.includes
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
 import play.twirl.api.HtmlFormat
-import views.html.CUIViews.SelfAssessmentCUIView
+import views.html.CUIViews.SelfAssessmentCUIDAv2View
 import views.html.pages.helpers.ViewSpecBase
 
 class sidebarSACUISpec extends ViewSpecBase with Matchers with AnyWordSpecLike {
 
-  private val viewWithTemplate = app.injector.instanceOf[SelfAssessmentCUIView]
+  private val viewWithTemplate = app.injector.instanceOf[SelfAssessmentCUIDAv2View]
 
   private def createView: () => HtmlFormat.Appendable = () => viewWithTemplate()(fakeRequest, messages)
